refactor:sortType 쿼리 동기화 로직을 훅/상수로 분리
