+++ src/widgets/listingsSection/index.ts
@@ -2,3 +2,4 @@ export * from "./ui/searchSection.tsx/searchSection";
 export * from "./ui/listingsSection";
 export * from "./ui/listingsCardDetailSection/listingsCardDetailSection";
 export * from "./listingsSectionPage";
+export * from "./model/useSyncSortTypeFromQuery";
