+++ src/widgets/listingsSection/model/model.ts
@@ -1,0 +1,2 @@
+export const DEADLINE_SORT_TYPE = "마감임박순";
+export const DEADLINE_QUERY_VALUE = "deadline";
